Regularize type-encoding of request input:
Avoid special treatment of query params-only and body-only requests. That is, always nest query params and body under an object field ("params" and "body", respectively), even if it's the only request input.

Consequently, don't need to look at the *value* of RequestSchema to understand the request format; it's all in the type.